migrated tests, removed empty tests
