--- packages/augur-ui/test/positions/actions/load-account-trades-test.js
@@ -1,1 +1,0 @@
-describe("modules/positions/actions/load-account-trades.js", () => {});
